test-all-crates: conditional options, to test more crates
This new feature lets us provide the "enable these options which are
needed to make the tests pass" featrure, which is different for each
of the afflicted crates.  Then we can test these crates
    tor-hsservice
    arti
    arti-client
which minimal features.

This will be important in a moment, as we're going to want to be
relying on actually minimal features tests in arti cfg.rs.