python: Decode JSON in rpc responses more aggressively
This tweaks the APIs to return a dict in cases where we know
that we are getting a result, and to return an ArtiRpcResponse
in cases where we don't know what kind of response we're getting.
It also expands our Error objects to support decoding.

Rationale: everybody who is using this libaray will want to
decode the json objects that they receive.  By doing it in the
library, we can save them some annoyance.