CI integration-chutney: install chutney outside of setup script
Currently just copy-paste from the integration-chutney-shadow job.
This will be simplified and/or folded into an external script pending
<https://gitlab.torproject.org/tpo/core/chutney/-/merge_requests/31>.

Note that this job is currently disabled, making this commit roughly a
no-op. This is just to proactively try to keep this version of the job
unbroken for if and when we reenable it.