tor-keymgr: Allow dead code in dummy KeyMgr.
The fields are never read but need to be present in order for the
dummy `KeyMgrBuilder` to have the same interface as the real one.

(We could of course manually implement the builder, or let the dummy
`KeyMgrBuilder` impl diverge from the real one, but for now let's leave
this as-is and simply mark the fields as unused).