PoW: Consolidate feature flags into a single flag.
This replaces the hs-pow, hs-pow-full, hs-pow-v1, pow-v1, and pow-full
features with a single hs-pow-full feature.

It's possible that in the future we will want to split different schemes
into different features, but we can do that when it comes up.

For now, having this as a single flag makes it clearer what's going on,
since the previous thing was not actually expressive enough to capture
some things we care about (like "at least one pow scheme is enabled"
that works in a future-compatible way).

This change is not semver breaking since it's a experimental feature.

Related: #1751