Define macros to create a map-builder type.
We're about to need this to define some of the configuration
for RPC listeners.

The provided map-builder type is much less voluminous in terms of
APIs added than the list-builder type: It just uses Deref* and sub_builder.
I think this simplicity may be a win.

I'll need to rethink how this type handles defaults in order to
implement connect points properly; I've left XXXXs for that issue.