tor-cert: Add experimental API for building EncodedEd25519Certs.
This will enable us to deserialize byte slices as `EncodedEd25519Certs`.

Needed because this type will be used to representing a parsed + validated
cert retrieved from the keystore.

Technically, we *could* do without this function by defining a separate
newtype wrapper over `Vec<u8>` to represent the validated cert data, but
IMO adding a second encoded ed25519 cert type in another crate might be
confusing later down the line (because the two types will be nearly
identical, and are bound to eventually diverge in terms of API and
implementation).

Part of #1137