RPC spec: specify banner format.
See #1753.

This banner message will give the client a clear signal that the
server is ready, so that the client will know that it can read
files (such as a cookie file) written by the server.

(We can't use "is the port bound" as a way to signal that the
server is ready, since the server needs to bind the port in order
to determine that there isn't another server running... which
it needs to do before it can write an rpc cookie file.)