circ: Specialize the circparams from netparams function
Congestion control parameters have specific values depending on the
circuit type. Instead of using a CircuitType, which is removed in this
commit, specialize the function in this case onion and exit.

This allows us to get rid of CircuitType and solely use TargetCircUsage
instead.

At this commit, we use .expect() on the Builder. Future commit will
remove this to return a Result in case of failure. Worth noting that we
don't expect one.

Signed-off-by: David Goulet <dgoulet@torproject.org>