rpc: Strengthen guarantees from Cancel.
Cancellation is now fallable, which allows us to detect attempts
to cancel which cannot work.

We now guarantee that when you try to cancel a `Cancel<F>` future,
either the cancel operation will succeed, or the future will return
(or will have already returned) Ok(), but not both, and not neither.

Closes #818.