rpclib: Avoid overuse of execute_internal_ok.
The execute_internal_ok method converts every error response into an
internal error; as such, it's only appropriate when there is no way
for a well-behaved Arti instance to give an error response.
But we had been using it in a few places where errors were possible
under other circumstances.

This commit fixes that behavior, and adds documentation to help
avoid it.