proto: Remove ConversationInHandler
It is unused but most importantly it allows any RELAY cell to be sent
from anywhere in the code which is really not desirable because it is
skipping congestion control.

It also allows us to remove the `control_tx` from the reactor which is
one less channel to track/understand/think about.

This opens up the door to all sorts of problems especially side channel
that can be exploited if we are not careful.

We can always bring this back if we need it but for now, it is unused
and allows us to remove the `CtrlMsg::SendRelayCell` control message.

No code behavior change.

Signed-off-by: David Goulet <dgoulet@torproject.org>