proto: Move the close stream action into CircHop
This is the first commit of a series to make CircHop own the code of its
related actions.

This commit moves the "close stream" code into CircHop now owning the
action and offloading the Reactor code.

In order to do this, and future changes like for example handling BEGIN,
the CircHop needs to be aware of its hop number. This change also allows
us to add the hop number to logging. It also copies the reactor unique
ID to each hop for logging purposes.

Future commit will continue to move more code into CircHop.

Signed-off-by: David Goulet <dgoulet@torproject.org>