relay-selection: Add FamilyRules to exclude_relays_in_same_family.
When we're trying to exclude relays by family,
we need to know which lists to look at.
This information ultimately comes from the network parameters.

We could avoid this change if we just told clients
"look at all family information all the time",
but that's not what the proposal says.

This is a breaking change.