tor-proto: Implement From<Bug> for ReactorError.
We will soon have a `ConfluxSet` type. Some of its operations will
return `Bug` (for example, the method for getting the *only* leg of the
conflux set will return a `Bug` if the set has no legs, or more than 1
leg).

This conversion function will make it easier these errors to
`ReactorError`.