Impose a maximum on our fallback estimated timeout
The fallback timeout is the one that we use when we have
insufficient data.  We reset our observations, and maybe rebuild
our circuits, when we find that too many circuits have failed
recently.  When we do so, we double our fallback timeout.

Previously we had no limit, which could lead to overflow (#1693).
In this commit we impose a maximum of 2 hours,
which is ridiculously high.
(C tor uses a maximum of INT32_MAX seconds, which is even more
ridiculously high.)

Closes #1693.